Revisions
Changed URL structure so that html extensions no longer show. For
example, you navigate to /certification/ instead of
/certification/certification.html

Removed duplicate CSS files and directed all link elements to the main
CSS file.

Chnaged navbar so that it is no longer fixed and lies below the added
header with the Green Office Program brand.